pyproject.toml
[project]
name = "spotify-stop-ai"
version = "0.1.0"
description = "Auto-skip AI-generated artists on Spotify using real-time monitoring and classification"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "spotipy>=2.23.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.9.0",
    "SPARQLWrapper>=2.0.0",
    "aiolimiter>=1.1.0",
    "duckduckgo-search>=6.3.0",
    "jinja2>=3.1.0",
]

[project.scripts]
spotify-stop-ai = "spotify_stop_ai.main:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
